requirements.txt
flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
mysqlclient
numpy
openpyxl
python-ldap
SQLAlchemy
wtforms_components
